align with protobuf by removing GetConsensusInfo and returning BidList and Provisioners from the ExecuteStateTransition RPC call
